feat: move Layout into own component
